+++ internals-js/src/operations.ts
@@ -132,6 +132,15 @@ export class Field<TArgs extends {[key: string]: any} = {[key: string]: any}> ex
     return newField;
   }
 
+  withUpdatedAlias(newAlias: string | undefined): Field<TArgs> {
+    const newField = new Field<TArgs>(this.definition, this.args, this.variableDefinitions, newAlias);
+    for (const directive of this.appliedDirectives) {
+      newField.applyDirective(directive.definition!, directive.arguments());
+    }
+    this.copyAttachementsTo(newField);
+    return newField;
+  }
+
   appliesTo(type: ObjectType | InterfaceType): boolean {
     const definition = type.field(this.name);
     return !!definition && this.selects(definition);
@@ -139,7 +148,7 @@ export class Field<TArgs extends {[key: string]: any} = {[key: string]: any}> ex
 
   selects(definition: FieldDefinition<any>, assumeValid: boolean = false): boolean {
     // We've already validated that the field selects the definition on which it was built.
-    if (definition == this.definition) {
+    if (definition === this.definition) {
       return true;
     }
 
@@ -292,10 +301,18 @@ export class FragmentElement extends AbstractOperationElement<FragmentElement> {
   }
 
   withUpdatedSourceType(newSourceType: CompositeType): FragmentElement {
+    return this.withUpdatedTypes(newSourceType, this.typeCondition);
+  }
+
+  withUpdatedCondition(newCondition: CompositeType | undefined): FragmentElement {
+    return this.withUpdatedTypes(this.sourceType, newCondition);
+  }
+
+  withUpdatedTypes(newSourceType: CompositeType, newCondition: CompositeType | undefined): FragmentElement {
     // Note that we pass the type-condition name instead of the type itself, to ensure that if `newSourceType` was from a different
     // schema (typically, the supergraph) than `this.sourceType` (typically, a subgraph), then the new condition uses the
     // definition of the proper schema (the supergraph in such cases, instead of the subgraph).
-    const newFragment = new FragmentElement(newSourceType, this.typeCondition?.name);
+    const newFragment = new FragmentElement(newSourceType, newCondition?.name);
     for (const directive of this.appliedDirectives) {
       newFragment.applyDirective(directive.definition!, directive.arguments());
     }
@@ -311,12 +328,18 @@ export class FragmentElement extends AbstractOperationElement<FragmentElement> {
     const fragmentParent = this.parentType;
     const typeCondition = this.typeCondition;
     if (selectionParent != fragmentParent) {
-      // As long as there an intersection between the type we cast into and the selection parent, it's ok.
-      validate(
-        !typeCondition || runtimeTypesIntersects(selectionParent, typeCondition),
-        () => `Cannot add fragment of parent type "${this.parentType}" to selection set of parent type "${selectionSet.parentType}"`
-      );
-      return this.withUpdatedSourceType(selectionParent);
+      // This usually imply that the fragment is not from the same sugraph than then selection. So we need
+      // to update the source type of the fragment, but also "rebase" the condition to the selection set
+      // schema.
+      let updatedTypeCondition: CompositeType | undefined = undefined;
+      if (typeCondition) {
+        const typeInSchema = selectionParent.schema().type(typeCondition.name);
+        validate(typeInSchema, () => `Cannot add ${this} to selection of parent type ${selectionParent}: cannot find condition type in schema of parent type`);
+        validate(isCompositeType(typeInSchema), () => `Cannot add ${this} to selection of parent type ${selectionParent}: condition type in schema is a ${typeInSchema.kind}`);
+        validate(runtimeTypesIntersects(selectionParent, typeInSchema), () => `Cannot add ${this} to selection of parent type ${selectionParent}: condition type in schema does not intersect ${selectionParent}`);
+        updatedTypeCondition = typeInSchema;
+      }
+      return this.withUpdatedTypes(selectionParent, updatedTypeCondition);
     }
     return this;
   }
@@ -967,6 +990,22 @@ export class SelectionSet extends Freezable<SelectionSet> {
     return this._cachedSelections;
   }
 
+  fieldsInSet(): { path: string[], field: FieldSelection, directParent: SelectionSet }[] {
+    const fields = new Array<{ path: string[], field: FieldSelection, directParent: SelectionSet }>();
+    for (const selection of this.selections()) {
+      if (selection.kind === 'FieldSelection') {
+        fields.push({ path: [], field: selection, directParent: this });
+      } else {
+        const condition = selection.element().typeCondition;
+        const header = condition ? [`... on ${condition}`] : [];
+        for (const { path, field, directParent } of selection.selectionSet.fieldsInSet()) {
+          fields.push({ path: header.concat(path), field, directParent });
+        }
+      }
+    }
+    return fields;
+  }
+
   usedVariables(): Variables {
     let variables: Variables = [];
     for (const byResponseName of this._selections.values()) {
@@ -1146,6 +1185,23 @@ export class SelectionSet extends Freezable<SelectionSet> {
     ++this._selectionCount;
     this._cachedSelections = undefined;
     return toAdd;
+  }
+
+  /**
+   * If this selection contains a selection of a field with provided response name at top level, removes it.
+   *
+   * @return whether a selection was removed.
+   */
+  removeTopLevelField(responseName: string): boolean {
+    // It's a bug to try to remove from a frozen selection set
+    assert(!this.isFrozen(), () => `Cannot remove from frozen selection: ${this}`);
+
+    const wasRemoved = this._selections.delete(responseName);
+    if (wasRemoved) {
+      --this._selectionCount;
+      this._cachedSelections = undefined;
+    }
+    return wasRemoved;
   }
 
   addPath(path: OperationPath, onPathEnd?: (finalSelectionSet: SelectionSet | undefined) => void) {
@@ -1437,6 +1493,10 @@ export class FieldSelection extends Freezable<FieldSelection> {
     this.selectionSet = isLeafType(type) ? undefined : (initialSelectionSet ? initialSelectionSet.cloneIfFrozen() : new SelectionSet(type as CompositeType));
   }
 
+  get parentType(): CompositeType {
+    return this.field.parentType;
+  }
+
   protected us(): FieldSelection {
     return this;
   }
@@ -1624,6 +1684,10 @@ export class FieldSelection extends Freezable<FieldSelection> {
     return new FieldSelection(this.field, newSubSelection);
   }
 
+  withUpdatedField(newField: Field<any>): FieldSelection {
+    return new FieldSelection(newField, this.selectionSet);
+  }
+
   equals(that: Selection): boolean {
     if (this === that) {
       return true;
@@ -1710,6 +1774,10 @@ export abstract class FragmentSelection extends Freezable<FragmentSelection> {
   abstract updateForAddingTo(selectionSet: SelectionSet): FragmentSelection;
 
   abstract withUpdatedSubSelection(newSubSelection: SelectionSet | undefined): FragmentSelection;
+
+  get parentType(): CompositeType {
+    return this.element().parentType;
+  }
 
   protected us(): FragmentSelection {
     return this;